move liftables
