+++ distage/distage-model/src/main/scala/com/github/pshirshov/izumi/distage/model/reflection/macros/WithDILiftables.scala
@@ -1,13 +1,11 @@
-package com.github.pshirshov.izumi.distage.model.definition.reflection
+package com.github.pshirshov.izumi.distage.model.reflection.macros
 
 import com.github.pshirshov.izumi.distage.model.reflection.universe.{RuntimeDIUniverse, StaticDIUniverse}
 
-class DIUniverseMacros[D <: StaticDIUniverse](val u: D) {
-  import u._
-  import u.u._
+trait WithDILiftables {
+  this: StaticDIUniverse =>
 
-  def modifiersForAnns(anns: List[Annotation]): Modifiers =
-    Modifiers.apply(NoFlags, typeNames.EMPTY, anns.map(_.tree))
+  import u._
 
   implicit val liftableRuntimeUniverse: Liftable[RuntimeDIUniverse.type] =
     { _: RuntimeDIUniverse.type => q"${symbolOf[RuntimeDIUniverse.type].asClass.module}" }
@@ -57,34 +55,50 @@ class DIUniverseMacros[D <: StaticDIUniverse](val u: D) {
     context => q"{ new $RuntimeDIUniverse.DependencyContext.ConstructorParameterContext(${context.definingClass}) }"
   }
 
+  implicit val liftableMethodParameterContext: Liftable[DependencyContext.MethodParameterContext] = {
+    context => q"{ new $RuntimeDIUniverse.DependencyContext.MethodParameterContext(${context.definingClass}, ${context.factoryMethod}) }"
+  }
+
+  implicit val liftableParameterContext: Liftable[DependencyContext.ParameterContext] = {
+    case context: DependencyContext.ConstructorParameterContext => q"$context"
+    case context: DependencyContext.MethodParameterContext => q"$context"
+  }
+
   // SymbolInfo
 
   implicit val liftableSymbolInfo: Liftable[SymbolInfo] = {
     info => q"""
-    { $RuntimeDIUniverse.SymbolInfo.StaticSymbol(${info.name}, ${info.finalResultType}, ${info.annotations}, ${info.isMethodSymbol}, ${info.definingClass}) }
+    { $RuntimeDIUniverse.SymbolInfo.StaticSymbol(${info.name}, ${info.finalResultType}, ${info.annotations}, ${info.definingClass}) }
        """
   }
+
+  // Associations
+
+  implicit val liftableParameter: Liftable[Association.Parameter] = {
+    case Association.Parameter(context, symb, wireWith) =>
+      q"{ new $RuntimeDIUniverse.Association.Parameter($context, $symb, $wireWith)}"
+    case Association.Parameter(context, _, _) =>
+      throw new RuntimeException(s"Expected ConstructorParameterContext but got $context")
+    }
 
   // Annotations
 
   case class TreeLiteral(tree: Tree)
 
   implicit val liftableLiteralTree: Liftable[TreeLiteral] = {
+    case TreeLiteral(Literal(c: Constant)) => q"""{
+      _root_scala.reflect.runtime.universe.Literal(_root_scala.reflect.runtime.universe.Constant($c))
+      }"""
     case TreeLiteral(tree) => q"""{
-    import _root_.scala.reflect.runtime.universe._
+      import _root_.scala.reflect.runtime.universe._
 
-    _root_.scala.StringContext(${showCode(tree, printRootPkg = true)}).q.apply()
-    }"""
+      _root_.scala.StringContext(${showCode(tree, printRootPkg = true)}).q.apply()
+      }"""
   }
 
   implicit val liftableAnnotation: Liftable[Annotation] = {
     case Annotation(tpe, args, _) => q"""{
-    _root_.scala.reflect.runtime.universe.Annotation.apply(${SafeType(tpe)}.tpe, ${args.map(TreeLiteral(_))}, _root_.scala.collection.immutable.ListMap.empty)
+    _root_.scala.reflect.runtime.universe.Annotation.apply(${SafeType(tpe)}.tpe, ${args.map(TreeLiteral)}, _root_.scala.collection.immutable.ListMap.empty)
     }"""
   }
-
-}
-
-object DIUniverseMacros {
-  def apply(u: StaticDIUniverse): DIUniverseMacros[u.type] = new DIUniverseMacros[u.type](u)
 }
